services/requirements.txt
Flask
kafka-python
flask_apscheduler
PyJWT==1.7.1
pycryptodome==3.10.4
